+++ frontend/src/app/pages/exercises/exercise/shared/exercise-map/feature-managers/map-images-feature-manager.ts
@@ -80,6 +80,9 @@ export class MapImageFeatureManager extends MoveableFeatureManager<MapImage> {
                 MapImagePopupComponent,
                 feature,
                 [feature.getId() as UUID],
+                [],
+                [],
+                [],
                 {
                     mapImageId: feature.getId() as UUID,
                 }