+++ frontend/src/app/pages/exercises/exercise/shared/exercise-map/feature-managers/simulated-region-feature-manager.ts
@@ -196,6 +196,9 @@ export class SimulatedRegionFeatureManager
             elementUUID: feature.getId()?.toString(),
             component: SimulatedRegionPopupComponent,
             closingUUIDs: [feature.getId() as UUID],
+            markedForParticipantUUIDs: [],
+            markedForTrainerUUIDs: [],
+            changedLayers: [],
             context: {
                 simulatedRegionId: feature.getId() as UUID,
             },